requirements.txt
fastapi==0.111.1
uvicorn==0.30.6
python-dotenv==1.0.1
ell-ai==0.0.11